\name{BA22022002-package}
\alias{BA22022002-package}
\alias{BA22022002}
\docType{package}
\title{
  Student BA22022002's R package for StatComp
}
\description{
  Functions & examples of R & Rcpp for illustration.
}
\author{
Liang Wu

Maintainer: Liang Wu <wul@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }

