% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA22022002.R
\name{pearson.r}
\alias{pearson.r}
\title{Pearson correlation}
\usage{
pearson.r(xs, ys)
}
\arguments{
\item{xs}{the first vector}

\item{ys}{the second vector}
}
\value{
Pearson correlation between xs and ys
}
\description{
Calculate pearson correlation between xs and ys
}
\examples{
\dontrun{
xs <- runif(100)
ys <- runif(100)
res <- pearson.r(xs, ys)
}
}
