% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rankdata}
\alias{rankdata}
\title{A function assign ranks to data}
\usage{
rankdata(xs)
}
\arguments{
\item{xs}{data to be assigned with rank}
}
\value{
Rank vector of input data
}
\description{
A function assign ranks to data
}
