% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BA22022002.R
\name{spearman.r}
\alias{spearman.r}
\title{Spearman correlation}
\usage{
spearman.r(xs, ys)
}
\arguments{
\item{xs}{the first vector}

\item{ys}{the second vector}
}
\value{
Spearman correlation between xs and ys
}
\description{
Calculate spearman correlation between xs and ys
}
\examples{
\dontrun{
xs <- runif(100)
ys <- runif(100)
res <- spearman.r(xs, ys)
}
}
